+++ app/actions/bundle.actions.js
@@ -1,5 +1,7 @@
 import traverse from 'traverse';
 import log from 'electron-log';
+import { List, Map } from 'immutable';
+import throttledQueue from 'throttled-queue';
 import { bundleConstants } from '../constants/bundle.constants';
 import { bundleService } from '../services/bundle.service';
 import { updateSearchResultsForBundleId } from '../actions/bundleFilter.actions';
@@ -35,6 +37,7 @@ export function updateBundle(bundleId) {
     try {
       const rawBundle = await bundleService.fetchById(bundleId);
       if (!bundleService.apiBundleHasMetadata(rawBundle)) {
+        console.log(`Skipping updateBundle for ${bundleId}`);
         return; // hasn't downloaded metadata yet. (don't expect to be in our list)
       }
       dispatch(updateOrAddBundle(rawBundle));
@@ -83,7 +86,9 @@ function updateOrAddBundle(rawBundle) {
       }
     } else if (rawBundle.mode === 'store') {
       dispatch(addBundle(bundle, rawBundle));
-      // console.log(`Added bundle ${bundleId} from ${context}`);
+      console.log(`Added bundle ${bundleId} from listenStorerChangeMode`);
+    } else {
+      console.log(`Skipped bundle ${bundleId} with mode ${rawBundle.mode}`);
     }
   };
 }
@@ -199,7 +204,7 @@ export function setupBundlesEventSource(authentication) {
     };
     const listeners = {
       'storer/execute_task': listenStorerExecuteTaskDownloadResources,
-      'storer/change_mode': (e) => listenStorerChangeMode(e, dispatch, getState),
+      'storer/change_mode': (e) => dispatch(listenStorerChangeMode(e)),
       'uploader/job': (e) => listenUploaderJob(e, dispatch, getState().bundles.uploadJobs),
       'uploader/createJob': (e) => listenUploaderCreateJob(e, dispatch),
       'downloader/receiver': listenDownloaderReceiver,
@@ -237,14 +242,16 @@ export function setupBundlesEventSource(authentication) {
     // console.log(e);
   }
 
-  function listenStorerChangeMode(e, dispatch) {
-    // console.log(e);
-    const data = JSON.parse(e.data);
-    const bundleId = data.args[0];
-    if (bundleId.startsWith('session')) {
-      return; // skip session change modes
-    }
-    dispatch(updateBundle(bundleId));
+  function listenStorerChangeMode(e) {
+    return dispatch => {
+      // console.log(e);
+      const data = JSON.parse(e.data);
+      const bundleId = data.args[0];
+      if (bundleId.startsWith('session')) {
+        return; // skip session change modes
+      }
+      dispatch(updateBundle(bundleId));
+    };
   }
 
   function listenStorerWriteResource(event) {
@@ -397,19 +404,18 @@ export function setupBundlesEventSource(authentication) {
   }
 }
 
+const throttleAddBundle = throttledQueue(4, 1000, true);
+
 function addBundle(bundle, rawBundle) {
-  return dispatch => {
+  return dispatch => throttleAddBundle(() => {
     dispatch({
       type: bundleConstants.ADD_BUNDLE,
       bundle,
-      rawBundle,
-      meta: {
-        throttle: true
-      }
+      rawBundle
     });
     dispatch(updateSearchResultsForBundleId(bundle.id));
     dispatch(removeExcessBundles());
-  };
+  });
 }
 
 function isInDraftMode(bundle) {
@@ -423,13 +429,14 @@ function removeExcessBundles() {
       // don't auto-remove when processing a lot of downloads (i.e. of metadata for new bundles)
       return;
     }
+    console.log('remove excess bundles');
     const { addedByBundleIds, items } = bundles;
-    const itemsByBundleIds = items.reduce((acc, bundle) => ({ ...acc, [bundle.id]: bundle }), {});
-    const itemsByParentIds = items.filter(b => b.parent).reduce((acc, bundle) =>
-      ({ ...acc, [bundle.parent.bundleId]: bundle }), {});
+    const itemsByBundleIds = List(items).map(bundle => bundle.id).toSet();
+    const itemsByParentIds = List(items).filter(b => b.parent).reduce((acc, bundle) =>
+      acc.set(bundle.parent.bundleId, bundle), Map());
     const itemsByDblId = items.reduce((acc, bundle) => ({ ...acc, [bundle.dblId]: bundle }), {});
     const bundleIdsToRemove = Object.keys(addedByBundleIds).filter(addedId => {
-      if (addedId in itemsByBundleIds) {
+      if (itemsByBundleIds.includes(addedId)) {
         return false;
       }
       const addedBundle = addedByBundleIds[addedId];
@@ -448,7 +455,7 @@ function removeExcessBundles() {
         }
       }
       // don't delete if is parent of item in draft mode
-      const itemDisplayed = itemsByParentIds[addedId];
+      const itemDisplayed = itemsByParentIds.get(addedId);
       if (itemDisplayed && isInDraftMode(itemDisplayed)) {
         return false;
       }
@@ -460,7 +467,7 @@ function removeExcessBundles() {
   };
   thunk.meta = {
     debounce: {
-      time: 2500,
+      time: 10500,
       key: 'BUNDLES_REMOVE_EXCESS_BUNDLES'
     }
   };