+++ app/store/configureStore.dev.js
@@ -3,7 +3,6 @@ import thunk from 'redux-thunk';
 import { createHashHistory } from 'history';
 import { routerMiddleware, routerActions } from 'react-router-redux';
 import { createLogger } from 'redux-logger';
-import throttleActions from 'redux-throttle';
 import createDebounce from 'redux-debounced';
 import rootReducer from '../reducers';
 import * as counterActions from '../actions/counter';
@@ -16,13 +15,6 @@ const configureStore = (initialState?: counterStateType) => {
   const middleware = [];
   const enhancers = [];
 
-  const defaultWait = 300;
-  const defaultThrottleOption = { // https://lodash.com/docs#throttle
-    leading: true,
-    trailing: true
-  };
-  const throttleMiddleWare = throttleActions(defaultWait, defaultThrottleOption);
-  middleware.push(throttleMiddleWare);
   middleware.push(createDebounce());
 
   // Thunk Middleware